+++ tests/index.spec.js
@@ -1,13 +1,13 @@
 const assert = require('assert').strict;
 const supertest = require('supertest');
 require('dotenv').config({ path: '../.env' });
-const app = require('../app');
+const app = require('../api/app');
 
 describe('Tests for the api', function () {
   it('gets all movie data by search query', function (done) {
     supertest(app)
       .get('/movies')
-      .query({ searchTerm: 'Avengers' })
+      .query({ searchTerm: 'Auto da' })
       .expect(200)
       .end(function (err, res) {
         if (err) return done(err);
